\name{crauttools-package}
\alias{crauttools-package}
\alias{crauttools}
\docType{package}
\title{\packageTitle{crauttools}}
\description{\packageDescription{crauttools}}
\section{Package Content}{\packageIndices{crauttools}}
\author{\packageAuthor{crauttools}}
\section{Maintainer}{\packageMaintainer{crauttools}}
\keyword{package}
