% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{download.gdrive}
\alias{download.gdrive}
\title{download.gdrive}
\usage{
download.gdrive(
  f_name = "gdrive.file",
  share_url = NULL,
  download_url = NULL,
  parent_dir = getwd(),
  content = FALSE,
  ...
)
}
\arguments{
\item{f_name}{(default = "gdrive.file")string for what the file should be downloaded as}

\item{share_url}{(default = NULL) the share url of the file to download (ignored if download url is supplied)}

\item{download_url}{(deafult = NULL) the actual download url of the file to download}

\item{parent_dir}{(default = getwd()) the full path for the parent directory for the file to be downloaded to}

\item{content}{(default = FALSE) flag to determine whether to return the content or the file name}

\item{...}{extra parameters that will be passed to utils::download.file}
}
\value{
string of the full file path of downloaded file (or content if the content flag is TRUE)
}
\description{
download file from gdrive
Look at this site for further details of the issue:
https://syncwithtech.blogspot.com/p/direct-download-link-generator.html
}
\examples{
TODO: add nice examples
}
