% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{get_file_id.gdrive}
\alias{get_file_id.gdrive}
\title{get_file_id.gdrive}
\usage{
get_file_id.gdrive(share_url = NULL)
}
\arguments{
\item{share_url}{a share url from gdrive}
}
\value{
the file id token as a string
}
\description{
Extracts the file token from a gdrive share link
}
